merge feat into main

+++ src/api/v1/routes/tag.py
@@ -1,17 +1,16 @@
-from typing import Annotated
+from typing import Annotated, Optional
 from uuid import UUID
 
 from fastapi import APIRouter, Depends, HTTPException, Path, Query, status
 
 from src.db.session import DBSession
-from src.domain.schemas.tag import TagCreate, TagUpdate
+from src.domain.schemas.tag import TagCreate, TagsWithFiltersGet, TagUpdate
 from src.services.auth import AuthService
 from src.services.tag import TagService
 
 router = APIRouter(
     prefix="/tags",
     tags=["tags"],
-    dependencies=[Depends(AuthService.access_jwt_required)],
 )
 
 
@@ -19,14 +18,23 @@ router = APIRouter(
     "/",
     status_code=status.HTTP_200_OK,
 )
-async def create_tag(db: DBSession, data: TagCreate):
-    _service = TagService(session=db)
+async def create_tag(
+    db: DBSession,
+    data: TagCreate,
+    is_moderator: Annotated[bool, Depends(AuthService.is_moderator)],
+    _: Annotated[str, Depends(AuthService.access_jwt_required)],
+):
+    if not is_moderator:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail="Access denied"
+        )
 
+    _service = TagService(session=db)
     response = await _service.create_tag(data=data)
     if not response:
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Could not create tag, it should have a unique name!",
+            detail="Could not create tag",
         )
     else:
         return response
@@ -42,7 +50,7 @@ async def get_tag(db: DBSession, tag_id: Annotated[UUID, Path()]):
     response = await _service.get_tag(tag_id=tag_id)
     if not response:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="tag not found"
+            status_code=status.HTTP_404_NOT_FOUND, detail="Tag not found"
         )
     else:
         return response
@@ -51,22 +59,44 @@ async def get_tag(db: DBSession, tag_id: Annotated[UUID, Path()]):
 @router.get("/", status_code=status.HTTP_200_OK)
 async def get_tags(
     db: DBSession,
-    page: int = Query(1),
+    limit: Optional[int] = Query(50, gt=0),
+    offset: Optional[int] = Query(1, gt=0),
 ):
-    pass
+    _service = TagService(session=db)
+    filters = TagsWithFiltersGet(
+        limit=limit,
+        offset=offset,
+    )
+    response = await _service.get_tags(filters=filters)
+    if not response:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Tags not found"
+        )
+    else:
+        return response
 
 
 @router.patch(
     "/{tag_id}/",
     status_code=status.HTTP_200_OK,
 )
-async def update_tag(db: DBSession, tag_id: Annotated[UUID, Path()], data: TagUpdate):
-    _service = TagService(session=db)
+async def update_tag(
+    db: DBSession,
+    tag_id: Annotated[UUID, Path()],
+    data: TagUpdate,
+    is_moderator: Annotated[bool, Depends(AuthService.is_moderator)],
+    _: Annotated[str, Depends(AuthService.access_jwt_required)],
+):
+    if not is_moderator:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail="Access denied"
+        )
 
+    _service = TagService(session=db)
     response = await _service.update_tag(tag_id=tag_id, data=data)
     if not response:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="tag not found"
+            status_code=status.HTTP_400_BAD_REQUEST, detail="Could not update tag"
         )
     else:
         return response
@@ -76,13 +106,22 @@ async def update_tag(db: DBSession, tag_id: Annotated[UUID, Path()], data: TagUp
     "/{tag_id}/",
     status_code=status.HTTP_200_OK,
 )
-async def delete_tag(db: DBSession, tag_id: Annotated[UUID, Path()]):
-    _service = TagService(session=db)
+async def delete_tag(
+    db: DBSession,
+    tag_id: Annotated[UUID, Path()],
+    is_moderator: Annotated[bool, Depends(AuthService.is_moderator)],
+    _: Annotated[str, Depends(AuthService.access_jwt_required)],
+):
+    if not is_moderator:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail="Access denied"
+        )
 
+    _service = TagService(session=db)
     response = await _service.delete_tag(tag_id=tag_id)
     if not response:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="tag not found"
+            status_code=status.HTTP_404_NOT_FOUND, detail="Tag not found"
         )
     else:
         return response
